refactor: unify pkg/js runtime API
