+++ pkg/js/eval_test.go
@@ -104,8 +104,8 @@ func TestEvaluate(t *testing.T) {
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
 			t.Parallel()
-			evaluator := NewEvaluator(tt.tools)
-			result := evaluator.Evaluate(t.Context(), tt.input, nil)
+			jsRuntime := New(nil, tt.tools)
+			result := jsRuntime.Evaluate(t.Context(), tt.input, nil)
 			assert.Equal(t, tt.expected, result)
 		})
 	}
