+++ pkg/js/expand_test.go
@@ -126,8 +126,8 @@ func TestExpand(t *testing.T) {
 
 			env := testEnvProvider(tt.envVars)
 
-			expander := NewJsExpander(&env)
-			result := expander.Expand(t.Context(), tt.commands, nil)
+			jsRuntime := New(&env, nil)
+			result := jsRuntime.Expand(t.Context(), tt.commands, nil)
 
 			assert.Equal(t, tt.expected, result)
 		})
@@ -141,8 +141,8 @@ func TestExpandMap(t *testing.T) {
 		"USER": "alice",
 	})
 
-	expander := NewJsExpander(&env)
-	result := expander.ExpandMap(t.Context(), map[string]string{
+	jsRuntime := New(&env, nil)
+	result := jsRuntime.ExpandMap(t.Context(), map[string]string{
 		"none":   "List all files",
 		"simple": "Say hello to ${env.USER}",
 	})
@@ -160,16 +160,16 @@ func TestExpandMap_Reuse(t *testing.T) {
 		"USER": "alice",
 	})
 
-	expander := NewJsExpander(&env)
+	jsRuntime := New(&env, nil)
 
-	result := expander.ExpandMap(t.Context(), map[string]string{
+	result := jsRuntime.ExpandMap(t.Context(), map[string]string{
 		"none": "List all files",
 	})
 	assert.Equal(t, map[string]string{
 		"none": "List all files",
 	}, result)
 
-	result = expander.ExpandMap(t.Context(), map[string]string{
+	result = jsRuntime.ExpandMap(t.Context(), map[string]string{
 		"simple": "Say hello to ${env.USER}",
 	})
 	assert.Equal(t, map[string]string{
@@ -182,8 +182,8 @@ func TestExpandMap_Empty(t *testing.T) {
 
 	env := testEnvProvider(map[string]string{})
 
-	expander := NewJsExpander(&env)
-	result := expander.ExpandMap(t.Context(), map[string]string{})
+	jsRuntime := New(&env, nil)
+	result := jsRuntime.ExpandMap(t.Context(), map[string]string{})
 
 	assert.Empty(t, result)
 }
@@ -234,8 +234,8 @@ func TestExpandString(t *testing.T) {
 			t.Parallel()
 
 			env := testEnvProvider(map[string]string{})
-			expander := NewJsExpander(&env)
-			result := expander.Expand(t.Context(), tt.template, tt.values)
+			jsRuntime := New(&env, nil)
+			result := jsRuntime.Expand(t.Context(), tt.template, tt.values)
 			assert.Equal(t, tt.expected, result)
 		})
 	}
